refactor: update structure file and folder
