+++ src/app/app.ts
@@ -3,45 +3,40 @@ import data from '../Data/data.json';
 import dataEn from '../Data/dataEn.json';
 import dataProject from '../Data/projectData.json';
 
-import Main from './pages/Main';
 import Page from './pages/Gallery/Page.ts';
 import Header from './components/Header/Header.ts';
 import Localization from './pages/Localization/Localization.ts';
-import FirstPage from './pages/MainPage/FirstPage.ts';
 import Footer from './components/Footer/Footer.ts';
-import NavPanel from './pages/MainPage/NavPanel.ts';
+import NavPanel from './components/NavPanel/NavPanel.ts';
 import ActiveLink from './pages/ActivePages/activePages.ts';
-import Anonsement from './pages/MainPage/Anonsement.ts';
 import NewsPage from './pages/News/News.ts';
+import BiographyComponent from './pages/Biography/BiographyComponent.ts';
+import HomePage from './pages/HomePage/HomePage.ts';
 
 export default class App {
   header: Header;
   footer: Footer;
-  firstPage: FirstPage;
-  mainPage: Main;
+  homePage: HomePage;
   page: Page;
   localization: Localization;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   selectedData: any;
   navPanel: NavPanel;
   activeLink: ActiveLink;
-  announcement: Anonsement;
+  biography: BiographyComponent;
   news: NewsPage;
 
   constructor() {
     this.header = new Header(document.body);
     this.page = new Page();
-    this.mainPage = new Main();
-    this.firstPage = new FirstPage();
+    this.homePage = new HomePage();
     this.footer = new Footer(document.body);
 
     this.header.render();
-    this.firstPage.changeLanguageFirstPage();
     this.page.renderProject(dataProject);
     this.footer.render();
+    this.biography = new BiographyComponent();
 
-    const announcementSection = document.querySelector('.slider-box') as HTMLElement;
-    this.announcement = new Anonsement(announcementSection);
 
     const newsSection = document.querySelector('.news-section') as HTMLElement;
     const paginationContainer = document.querySelector('.pagination-container') as HTMLElement;
@@ -73,9 +68,12 @@ export default class App {
     return galleryPages.some(id => window.location.pathname.endsWith(`${id}.html`));
   }
 
-  // eslint-disable-next-line class-methods-use-this
   isNewsPage(): boolean {
     return window.location.pathname.endsWith('news.html');
+  }
+
+  isBioPage(): boolean {
+    return window.location.pathname.endsWith('biography.html');
   }
 
   renderGalleryPage() {
@@ -101,13 +99,19 @@ export default class App {
     this.news.render();
   }
 
+  renderBiography() {
+    this.biography.render();
+  }
+
   render() {
     if (this.isNewsPage()) {
       this.renderNewsPage();
     } else if (this.isGalleryPage()) {
       this.renderGalleryPage();
+    } else if (this.isBioPage()) {
+      this.renderBiography();
     }
 
-    this.announcement.render();
+    this.homePage.render();
   }
 }
