+++ src/app/interfaces/Event.ts
@@ -12,7 +12,7 @@ export interface EventProps {
   eventType: string;
   eventName: string;
   eventDate: string;
-  eventLink: string;
+  eventLink?: string;
   eventLocation: EventLocation;
 }
