+++ src/app/pages/HomePage/Components/CollageComponent.ts
@@ -1,0 +1,13 @@
+import Collage from '../../Collage/Collage';
+
+export default class CollageComponent {
+  collage: Collage;
+
+  constructor() {
+    this.collage = new Collage();
+  }
+
+  render() {
+    this.collage.collageAnimation();
+  }
+}
